Support up to 6 tupled values
Adding new rules like this obviously doesn’t scale. If we were using
an inductive structure (such as an `HList`) instead of flat tuples,
the rules would be much more simpler (and not limited to an arbitrary
arity!). Hopefully all of that will be simplified in Scala 3 where flat
tuples are unified with HLists.