Merge JSON guide and JsonEntities documentation
Another pass on the documentation. I think that, generally, the documentation does not clearly answer “how to” questions. This change aims at addressing this issue in the case of the `JsonEntities` documentation: instead of listing the operations of that algebra, we explain a bit more how to describe endpoints with JSON entities in various situations.