Specify behavior of `optField`
Fixes #123

- Decoders of optional fields successfully decode `None` if the field is absent from the JSON object or if it is present with the value `null`,
- Decoders of optional fields fail if the field is present but has an invalid value,
- Encoders of optional fields can omit the field or emit it with value `null`.

I was not sure whether it was a good idea or not to successfully decode fields with value `null` as `None`. As long as we don’t give another meaning to `null` this seems safe to me. Since it seems to be a common practice to use fields with value `null` to model the absence of value, I thought that was important that we support it.

The tests for the behavior of optional fields have been factored in a generic class that should be used (as a set of law) to validate any codec interpreter of `JsonSchemas`.