Option to configure OpenAPI schemas for coproducts
Introduces an overrideable setting in `openapi.JsonSchemas` for
configuring the encoding of OpenAPI schemas for coproducts. The
two options that are supported are:

  1. use `oneOf` in the base type to refer to all of the variants

  2. in addition to using `oneOf` in the base type to refer to all of
     the variants, the variants refer back to the base type with `allOf`

Alternative 1 leads to simpler schemas, but alternative 2 allows users
to work around some OpenAPI code generators' buggy handling of `oneOf`.
See the discussion in #135 for more on the tradeoffs.

Fixes #503