Add `title` annotation
The new `title` annotation sets the title of a schema in the generic
derivation process. This is likely to be more useful than the `name`
annotation, since the latter is actually the internal OpenAPI name,
which has to be a valid URI suffix.

The `name` annotation;s scaladoc now includes a note about OpenAPI's
use and a reference to `title`.