Add `NoDocsJsonSchemas` for decoder `JsonSchemas`
The new `NoDocsJsonSchemas` trait implements all the documentation
related methods of `JsonSchemas` as no-ops. This reduces the boilerplate
in decoder inteprereters: they can mix in `algebra.NoDocsJsonSchemas`
instead of having to implement a bunch of pass-through documentation
methods.

This factors out a good chunk of common code in the Circe, Play, and
Ujson interpreters.

Fixes #514.