Add `plainTextRequest` and clarify `textRequest`
The newly added `plainTextRequest` is a request that must have type
`text/plain`. The semantics of `textRequest` have also been clarfied.
The important bits are:

  * it can have any content type
  * if the content-type specifies some charset, that charset gets used

The server implementations of `textRequest` have been adjusted to all
match these semantics - it turned out that the servers were all doing
different things!

  * Akka-HTTP was doing the "right" thing
  * http4s was always decoding the body as UTF-8, regardless of charset
  * Play was not accepting content-types different that `text/plain`

The server semantics of `testRequest`/`plainTextRequest` are tested in
the new `TextEntities` test suite.