OpenAPI: use `allOf` with `$ref` whenever appropriate
`$ref` JSON schemas are a bit special in that they aren't supposed to
have any sibling properties at all. Properties that would normally be
fields in the JSON schema object are supposed to be unified with the
`$ref` schema object with an explicit `allOf`.

Setting examples, descriptions, etc. on a reference schema now behaves
in the intuitive way (at least intuitive for me): it will override the
example, description, etc. on the _local instance_ of the schema, but
it won't otherwise modify the undderlying schema.