Ship reusable testing code in testkit modules
The tests of the algebra module don’t test anything. Instead, they provide (abstract) test specifications to be used by the interpreters to validate that they conform to the specification, or fixtures that can be conveniently reused by other modules.

By moving these testkits into separate modules, it becomes easier for third parties to validate that their interpreters are correct.

It also simplifies the build definition because we don’t need anymore dependencies like `foo % "test->test;compile->compile"`.

Last, the test modules were also used to carry snippets of code used in the documentation. Those are not included in the testkit modules (and they have been duplicated in each interpreter). The idea is to have decoupled documentation for interpreters.